requirements.txt
flask
requests
azure-storage-blob